requirements.txt
asgiref==3.7.2
black==24.3.0
click==8.1.7
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
